requirements.txt
decorator==4.0.9
Django==1.9.4
django-appconf==1.0.1
django-compressor==2.0
django-tinymce==2.3.0
gunicorn==19.4.5
ipython==4.1.2
ipython-genutils==0.1.0
path.py==8.1.2
pexpect==4.0.1
pickleshare==0.6
psycopg2==2.6.1
ptyprocess==0.5.1
rcssmin==1.0.6
rjsmin==1.0.12
simplegeneric==0.8.1
six==1.10.0
traitlets==4.2.1
pillow==3.1.1
